pyproject.toml
[project]
name = "holoanalytics-mini2"
version = "0.1.0"
description = ""
authors = [
    {name = "tatto"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client (>=2.173.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (>=1.38.41,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "dbt-mysql (>=1.7.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
